requirements.txt
aiohttp==3.7.4.post0
asgiref==3.6.0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.5
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
fontawesomefree==6.2.1
idna==3.2
multidict==5.1.0
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.2.2
pandas==1.3.2
protobuf==4.21.11
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
